runner: split Lab offload into focused service modules (#4304)
Carve the 3,186-line `src/core/runner/lab.rs` into focused submodules so each
Lab offload concern stays close to its tests, without changing CLI behavior:

- `core::runner::lab::offload` keeps request execution, runner selection,
  workspace-sync-mode policy, trace-target git fetch refs, the
  `execute_lab_offload` facade entry, and the fallback helpers.
- `core::runner::lab::agent_task_bridge` owns inline AgentTask plan
  materialization, run-plan lifecycle mirroring, dispatch-envelope parsers,
  and the pre-dispatch failure helpers.
- `core::runner::lab::secrets` owns command-specific secret env hydration for
  agent-task and trace commands.
- `core::runner::lab::evidence` owns terminal Lab run evidence discovery for
  the run idempotency guard.
- `core::runner::lab::args_util` is a tiny shared argv inspection helper used
  by the other submodules.

`src/core/runner/lab.rs` shrinks from 3,186 lines to a ~37-line module shell
that wires up the submodules and re-exports the public Lab offload surface
(`execute_lab_offload`, `LabOffloadRequest`, `LabOffloadCommand`,
`LabOffloadOutcome`, `LabOffloadWorkspaceModePolicy`,
`LabRunnerSelectionSource`). `core::runners::execute_lab_offload` remains the
public facade entry point; nothing outside `runner` reaches into the new
submodules.

Tests move with the code they exercise, so each submodule now carries its own
`#[cfg(test)] mod tests` block; `lab_arg_tests` switches from `use super::*;`
to an explicit `lab_args` import now that `lab.rs` no longer re-exports them.

Verification:
- `cargo fmt --check`
- `cargo test git_materialization` (6 passed)
- `cargo test 'core::runner::lab'` (102 passed)

Closes #4304

## AI assistance
Tool OpenCode (openai/gpt-5.5), used for refactor implementation and tests.
